+++ 5_LT_companies_cars/main.py
@@ -1,5 +1,5 @@
 import logging
-from src.pipeline9 import run
+from src.pipeline import run
 
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
