+++ 5_LT_companies_cars/src/pipeline.py
@@ -1,5 +1,5 @@
 import datetime as dt
-import random
+import os
 import pandas as pd
 import io
 import logging
@@ -13,122 +13,123 @@ from pydantic import BaseModel
 from typing import List
 
 from src.mappings import file_configurations 
+
 # Configuration
 bucket_name = "lithuania_statistics"
 staging_location = f"gs://{bucket_name}/staging"
 temp_location = f"gs://{bucket_name}/temp"
 
+# Data classes
 class TableSchema(BaseModel):
     name: str
     data_type: str
 
 class UploadConfig(BaseModel):
     bucket_name: str
-    folder_name: str
     file_name: str
-    dataset_name: str
     table_name: str
     table_schema: List[TableSchema]
+
 
 class DownloadSave(beam.DoFn):
     def __init__(self):
         self.current_year = dt.date.today().year
         self.file_configurations = file_configurations
 
-    def download_file(self, zip_file_url):
-        header = {
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:98.0) Gecko/20100101 Firefox/98.0",
-        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,*/*;q=0.8",
-        "Accept-Language": "en-US,en;q=0.5",
-        "Accept-Encoding": "gzip, deflate",
-        "Connection": "keep-alive",
-        "Upgrade-Insecure-Requests": "1",
-        "Sec-Fetch-Dest": "document",
-        "Sec-Fetch-Mode": "navigate",
-        "Sec-Fetch-Site": "none",
-        "Sec-Fetch-User": "?1",
-        "Cache-Control": "max-age=0",
-    }
-        url = zip_file_url
-        local_filename = url.split('/')[-1]
+    def download_file(self, url):
+        """Download a file from the web and save it to GCS."""
+        # Check if 'sodra' is present in the URL
+        if 'sodra' in url:
+            logging.info(f'Skipping download for URL containing "sodra": {url}')
+            return None
+
+        header = {'User-Agent': 'Mozilla/5.0 (Windows NT 6.1; Win64; x64)'}
+        local_filename = url.split('/')[-1].replace("-", "_")
         with requests.get(url, headers=header) as r:
-            with open(local_filename, 'wb') as f:
-                f.write(r.content)
-                logging.info(f"File {local_filename} downloaded")
-        return local_filename
+            try:
+                with open(local_filename, 'wb') as f:
+                    f.write(r.content)
+                    logging.info(f'Downloaded {url} to {local_filename}, size in MB: {round(os.path.getsize(local_filename) / 1024 / 1024, 2)}')
+                return local_filename
+            except Exception as e:
+                logging.error(f"Error downloading file: {str(e)}")
+                return None
     
     def extract_zip_contents(self, zip_file):
-        # Extract the contents of the zip file.
-        extracted_files = {}
-        with zipfile.ZipFile(zip_file, "r") as zip_ref:
-            for file_name in zip_ref.namelist():
-                with zip_ref.open(file_name) as f:
-                    extracted_files[file_name] = f.read()
-        return extracted_files
-    
-    def download_and_extract(self, zip_file_url):
-        # Download the file and handle any download errors.
-        local_filename = self.download_file(zip_file_url)
-        if local_filename is None:
-            # Handle download failure here
+        """Extract the contents of a zip file."""
+        try:
+            extracted_files = {}
+            with open(zip_file, "rb") as zip_file_obj:
+                with zipfile.ZipFile(zip_file_obj) as zip_ref:
+                    for file_info in zip_ref.infolist():
+                        file_name = file_info.filename
+                        extracted_files[file_name] = zip_ref.read(file_info)
+                        logging.info(f'Extracted {file_name}')
+            return extracted_files
+        except Exception as e:
+            logging.error(f"Error extracting zip file, name: {zip_file}, error: {str(e)}")
             return {}
-
-        # Continue with extracting and processing the file.
-        extracted_files = {}
-        if local_filename.endswith(".zip"):
-            try:
-                with open(local_filename, "rb") as f:
-                    extracted_files = self.extract_zip_contents(f)
-            except Exception as e:
-                logging.error(f"Failed to extract ZIP file: {e}, {local_filename}")
-
-        return extracted_files
-    
-    def upload_files_to_bucket(self, extracted_files):
-        # Upload the files to the bucket.
-        client = storage.Client()
-        bucket = client.get_bucket(bucket_name)
-        for file_name, file_contents in extracted_files.items():
-            blob = bucket.blob(f"{bucket_name}/{file_name}")
-            blob.upload_from_string(file_contents)
-            logging.info(f"File {file_name} uploaded to bucket")
 
     def process(self, element):
         # Iterate over the file configurations.
         for file_configuration in self.file_configurations:
             # Get the file configuration.
-            file_name = file_configuration["file_name"]
             url = file_configuration["url"]
-            # Download and extract the file if it is a zip file.
-            if url.endswith(".zip"):
-                extracted_files = self.download_and_extract(url)
-            # Upload the file to the bucket.
-            self.upload_files_to_bucket(extracted_files)
-            yield file_name
+            local_filename = self.download_file(url)
+            extracted_files = self.extract_zip_contents(local_filename)
+            for file_name, file_contents in extracted_files.items():
+                # Upload the file to GCS.
+                bucket = storage.Client().bucket(bucket_name)
+                blob = bucket.blob(f'{bucket_name}/{file_configuration["file_name"]}')
+                blob.upload_from_string(file_contents, content_type='text/csv')
+                logging.info(f'Uploaded {file_name} to {bucket_name}/{file_configuration["file_name"]}')
+                # Delete the local file.
+                os.remove(local_filename)
 
+        #unzip all files in bucket endind with .zip
+        for blob in storage.Client().list_blobs(bucket_name):
+            if blob.name.endswith(".zip"):
+                blob.download_to_filename("/tmp/temp.zip")
+                with zipfile.ZipFile("/tmp/temp.zip", 'r') as zip_ref:
+                    zip_ref.extractall("/tmp/")
+                for file in os.listdir("/tmp/"):
+                    if file.endswith(".csv"):
+                        blob = bucket.blob(f'{bucket_name}/{file}')
+                        blob.upload_from_filename(f"/tmp/{file}")
+                        logging.info(f'Uploaded {file} to {bucket_name}/{file}')
+                        os.remove(f"/tmp/{file}")
+                os.remove("/tmp/temp.zip")
+                
+        #rename in bucket Sodra file
+        for blob in storage.Client().list_blobs(bucket_name):
+            if 'monthly' in blob.name and blob.name.endswith(".csv"):
+                new_name = file_configuration["file_name"]
+                new_blob = bucket.rename_blob(blob, f'{bucket_name}/{new_name}')
+                logging.info(f'File {blob.name} has been renamed to {new_blob.name}')
+
+        
 class UploadToBigQuery(beam.DoFn):
-    def __init__(self, config: UploadConfig = None):
-        self.config = config
-        self.storage_client = storage.Client()
-        self.bigquery_client = bigquery.Client()
+    def __init__(self):
         self.file_configurations = file_configurations
         self.transform_functions = {
             "Atviri_JTP_parko_duomenys.csv": self.transform_data_companies,
             "employees_salaries_raw.csv": self.transform_data_employees
         }
 
-    def read_file_from_gcs(self):
-        bucket = self.storage_client.bucket(self.config.bucket_name)
-        blob = bucket.blob(f'{self.config.folder_name}/{self.config.file_name}')
-        blob_as_string = blob.download_as_string()
-        delimiter = self.get_delimiter()
-        data_frame = pd.read_csv(io.BytesIO(blob_as_string), sep=delimiter, encoding='utf-8')
-        transform_function = self.transform_functions[self.config.file_name]
-        return transform_function(data_frame)
+    def read_file_from_gcs(self, config):
+        """Reads a file from GCS and returns a DataFrame."""
+        bucket = storage.Client().bucket(config.bucket_name)
+        blob = bucket.blob(f'{config.bucket_name}/{config.file_name}')
+        file_bytes = blob.download_as_bytes()
+        data_frame = pd.read_csv(io.BytesIO(file_bytes), sep=self.get_delimiter())
+        logging.info(f'Read {config.file_name} from GCS')
+        return self.transform_functions[config.file_name](data_frame)
     
     def get_delimiter(self):
+        """Returns a delimiter for a file."""
         for file_configuration in file_configurations:
             if file_configuration["file_name"] == self.config.file_name:
+                logging.info(f'Got delimiter for {self.config.file_name}')
                 return file_configuration["delimiter"]
         logging.error(f"Could not find delimiter for {self.config.file_name}")
         return None
@@ -140,6 +141,7 @@ class UploadToBigQuery(beam.DoFn):
         data_frame = data_frame.dropna(subset=['KOMERCINIS_PAV'])
         # make column KODAS as first column
         data_frame = data_frame[['KODAS'] + [col for col in data_frame.columns if col != 'KODAS']]
+        logging.info(f'Transformed {self.config.file_name}')
         return data_frame
     
     def transform_data_employees(self, data_frame: pd.DataFrame) -> pd.DataFrame:
@@ -162,16 +164,8 @@ class UploadToBigQuery(beam.DoFn):
             "Vidutinis darbo užmokestis (avgWage)": "vidutinis_darbo_uzmokestis",
             "Apdraustųjų skaičius (numInsured)": "apdraustuju_skaicius"
         })
+        logging.info(f'Transformed {self.config.file_name}')
         return data_frame
-
-    def upload_file_to_bigquery(self, data_frame: pd.DataFrame):
-        """Upload data to a BigQuery table."""
-        dataset = self.bigquery_client.dataset(self.config.dataset_name)
-        table = dataset.table(self.config.table_name)
-        job_config = self.get_job_config()
-        job = self.bigquery_client.load_table_from_dataframe(data_frame, table, job_config=job_config)
-        job.result()
-        logging.info(f'Uploaded {self.config.file_name} to {self.config.dataset_name}.{self.config.table_name}')
 
     def get_job_config(self) -> bigquery.LoadJobConfig:
         """Returns a LoadJobConfig for BigQuery table loading."""
@@ -181,6 +175,7 @@ class UploadToBigQuery(beam.DoFn):
         job_config.autodetect = True
         job_config.schema = self.get_table_schema()
         job_config.write_disposition = bigquery.WriteDisposition().WRITE_TRUNCATE
+        logging.info(f'Got job config for {self.config.file_name}')
         return job_config
 
     def get_table_schema(self) -> List[bigquery.SchemaField]:
@@ -188,28 +183,29 @@ class UploadToBigQuery(beam.DoFn):
         table_schema = []
         for field in self.config.table_schema:
             table_schema.append(bigquery.SchemaField(field.name, field.data_type))
+        logging.info(f'Got table schema for {self.config.file_name}')
         return table_schema
 
+    def upload_file_to_bigquery(self, config, data_frame, bigquery_client):
+        """Upload data to a BigQuery table."""
+        table = f'{bucket_name}.{config.table_name}'
+        job_config = self.get_job_config()
+        job = bigquery_client.load_table_from_dataframe(data_frame, table, job_config=job_config)
+        job.result()
+        logging.info(f'Uploaded {config.file_name} to {bucket_name}.{config.table_name}')
+
     def process(self, element):
-        # Iterate over the file configurations.
         for file_configuration in self.file_configurations:
-            # Get the file configuration.
-            upload_config = UploadConfig(
-                bucket_name = "lithuania_statistics",
-                folder_name = "companies_cars",
-                file_name = file_configuration["file_name"],
-                dataset_name = "lithuania_statistics",
-                table_name = file_configuration["table_name"],
-                table_schema = file_configuration["table_schema"]
+            config = UploadConfig(
+                bucket_name=bucket_name,
+                file_name=file_configuration["file_name"],
+                table_name=file_configuration["table_name"],
+                table_schema=file_configuration["table_schema"]
             )
-            uploader = UploadToBigQuery(upload_config)
-            data_frame = uploader.read_file_from_gcs()
-            uploader.upload_file_to_bigquery(data_frame)
-    
-
-# DownloadSave().process(None)
-# UploadToBigQuery().process(None)
-
+            self.config = config
+            data_frame = self.read_file_from_gcs(config)
+            bigquery_client = bigquery.Client()
+            self.upload_file_to_bigquery(config, data_frame, bigquery_client)
 
 def run():
     pipeline_options = PipelineOptions()
@@ -219,12 +215,18 @@ def run():
     pipeline_options.view_as(GoogleCloudOptions).staging_location = staging_location
     pipeline_options.view_as(GoogleCloudOptions).temp_location = temp_location
 
-    with beam.Pipeline(options=pipeline_options) as pipeline:
-        (
-            pipeline
-            | "Create" >> beam.Create([None])
-            | "Download and save" >> beam.ParDo(DownloadSave())
-            # | "Upload to BigQuery" >> beam.ParDo(UploadToBigQuery()
+    with beam.Pipeline(options=pipeline_options) as p:
+        # Download and save files from the web to GCS.
+        download_save = (
+            p
+            | 'Create' >> beam.Create([None])
+            | 'Download and save' >> beam.ParDo(DownloadSave())
+        )
+        # Upload files from GCS to BigQuery.
+        upload_to_bigquery = (
+            p
+            | 'Create again' >> beam.Create([None])
+            | 'Upload to BigQuery' >> beam.ParDo(UploadToBigQuery())
         )
